@@ -51,21 +51,21 @@
   ],
   "dependencies": {
     "arrify": "^1.0.0",
-    "babel-eslint": "^4.1.3",
+    "babel-eslint": "^4.1.5",
     "deep-assign": "^1.0.0",
-    "eslint": "^1.6.0",
-    "eslint-config-marlint": "^1.3.1",
+    "eslint": "^1.9.0",
+    "eslint-config-marlint": "^1.4.1",
     "eslint-plugin-babel": "^2.1.1",
-    "eslint-plugin-react": "^3.5.1",
+    "eslint-plugin-react": "^3.8.0",
     "get-stdin": "^5.0.0",
-    "globby": "^3.0.0",
-    "meow": "^3.4.2",
+    "globby": "^3.0.1",
+    "meow": "^3.5.0",
     "object-assign": "^4.0.1",
     "pkg-conf": "^1.0.1",
     "update-notifier": "^0.5.0"
   },
   "devDependencies": {
     "ava": "^0.2.0",
-    "marlint": "traveloka/marlint#v1.3.1"
+    "marlint": "traveloka/marlint#v1.4.0"
   }
 }
